Refactoring this abomination
